basic rockets setup
