+++ src/App.js
@@ -4,18 +4,10 @@ import {
   Route,
   RouterProvider,
 } from 'react-router-dom';
-/* eslint-disable */
-
-// import RootLayout from "./layout/RootLayout";
-// import MissionsPage from "./pages/MissionsPage";
-// import MyProfilePage from "./Pages/MyProfilePage";
-// import RocketsPage from "./Pages/RocketsPage";
 import RootLayout from './layout/RootLayout';
-// import MissionsPage from "./Pages/MissionPage";
-// import RocketsPage from "./"
-import RocketsPage from './Page/RocketsPage';
-import MissionsPage from './Page/MissionPage.js';
-import MyProfilePage from './Page/MyProfilePage.js';
+import MissionsPage from './pages/MissionsPage';
+import MyProfilePage from './pages/MyProfilePage';
+import RocketsPage from './pages/RocketsPage';
 
 const router = createBrowserRouter(
   createRoutesFromElements(
